First working version.
Beware, here we couldn't test if importing from another project works,
because of an issue with submodules+restrictions on source files
that must not be out of source.
This cause a problem if a test is setup to be it's
own sub-project, like what `bdep new -t lib` generates by default.